Long타입 변경

+++ src/main/java/project/fahiru/domain/CookDetail.java
@@ -21,7 +21,7 @@ public class CookDetail extends TrendyFoodCook{
 
     private String cookTime;
 
-    private int cookAmount;
+    private Long cookAmount;
 
     private String openAccept;
